+++ CHANGELOG.md
@@ -3,12 +3,11 @@ All notable changes to this project will be documented in this file.
 
 ## [1.1.0] - 2017-06-03
 ### Changed
-- Fix bad behaviour on base processor
-- Fix import/attributeValueProcessor and import/familyProcessor to avoid useless Writer call 
+- Behaviour on base processor
+- import/attributeValueProcessor, import/familyProcessor, import/productProcessor to avoid useless Writer call 
 
 ### Fixed
 - Fix bad json formatting in locales/en.js and locales/fr.js files
-- Update outdated unreleased diff link.
 
 ### Added
 - Some specification to the README.md for import functionnality